% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nozero.R
\name{nozero}
\alias{nozero}
\title{No zero before decimal}
\usage{
nozero(x)
}
\arguments{
\item{x}{A number < 1}
}
\value{
a string without quotes with the appropriately formatted p value.
}
\description{
Truncates a decimal to three decimal places
}
\details{
Truncates a < 1 number to three decimal places with no 0 before the decmial places, for use in RMarkdown inline code for APA style documents.
}
