% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.p.R
\name{report.p}
\alias{report.p}
\title{Report p}
\usage{
report.p(p)
}
\arguments{
\item{p}{A numeric p value.}
}
\value{
a string without quotes with the appropriately formatted p value.
}
\description{
Formats p values appropriately for RMarkdown documents.
}
\details{
Formats p values appropriately for Rmarkdown documents in APA style by returning the p value as > .001 or = .xxx to be used for inline code reporting of p values.
}
