% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termcounter.R
\name{termcount}
\alias{termcount}
\title{termcount}
\usage{
termcount(x)
}
\arguments{
\item{x}{A vector of strings to be parsed.}
}
\value{
A corresponding numeric vector of number of terms in the string
}
\description{
Counts the number of distinct terms used in a a string.
Counts the number of terms used in a string that are separated by semicolon or comma (if semicolons were not used).
}
