% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varname_label.R
\name{varname_label}
\alias{varname_label}
\title{varname_label
 
 Renames variables based on text parsed from a data key.
 
 Renames variables by parsing text from descriptions of the variable. 
 Must be in the format of a data frame (named datakey) with variables matching the data set (named data), and containing only a single string.
 Intended to work with Qualtrics' output of question text. 
 
 @param vargrp String pattern to rename, renames all variables with this pattern in them.
 @param label A string common label to preface the new variable name with. 
 @param pattern A string section to be removed from labels in the data key, such that any remaining text is added to the variable name.
 @param key True returns the renamed data key, false returns the data. Defaults to false.
 @export}
\usage{
varname_label(vargrp, label, pattern, key = FALSE)
}
\description{
varname_label
 
 Renames variables based on text parsed from a data key.
 
 Renames variables by parsing text from descriptions of the variable. 
 Must be in the format of a data frame (named datakey) with variables matching the data set (named data), and containing only a single string.
 Intended to work with Qualtrics' output of question text. 
 
 @param vargrp String pattern to rename, renames all variables with this pattern in them.
 @param label A string common label to preface the new variable name with. 
 @param pattern A string section to be removed from labels in the data key, such that any remaining text is added to the variable name.
 @param key True returns the renamed data key, false returns the data. Defaults to false.
 @export
}
